no message
git-svn-id: http://svn.code.sf.net/p/gpac/code/trunk/gpac@79 63c20433-aa62-49bd-875c-5a186b69a8fb

+++ modules/aac_in/faad_dec.c
@@ -59,7 +59,7 @@ static GF_Err FAAD_AttachStream(GF_BaseDecoder *ifcg, u16 ES_ID, unsigned char *
 	ctx->codec = faacDecOpen();
 	if (!ctx->codec) return GF_IO_ERR;
 
-	if (faacDecInit2(ctx->codec, (unsigned char *) decSpecInfo, decSpecInfoSize, (unsigned long *) &ctx->sample_rate, (u8 *) &ctx->num_channels) < 0)
+	if ( (s8) faacDecInit2(ctx->codec, (unsigned char *) decSpecInfo, decSpecInfoSize, (unsigned long *) &ctx->sample_rate, (u8 *) &ctx->num_channels) < 0)
 		return GF_NON_COMPLIANT_BITSTREAM;
 
 	gf_m4a_get_config((unsigned char *) decSpecInfo, decSpecInfoSize, &a_cfg);
